test(e2e-tests): update publish horizon glue codes
